Revert peer dependency upgrade
I increased this when I saw that React Hooks Testing Library
peer-depends on React >=16.9.0, but I just realized that this is a sort
of breaking change for my package. Nothing in the package itself depends
on any new features in v16.9, and React Hooks Testing Library is only
used in development, where the dev dependency version range is more than
sufficient.